+++ src/types/responses/accounts.ts
@@ -12,3 +12,4 @@ export type AccountAssets = {
 export type AccountAddresses = OpenApiResponseTypes['account_addresses_content'];
 
 export type AccountAddressesTotal = OpenApiResponseTypes['account_addresses_total'];
+export type AccountUtxos = OpenApiResponseTypes['account_utxo_content'];